refactor: implementing new register page
