+++ src/app/(smal)/auth/register/page.tsx
@@ -1,0 +1,28 @@
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle
+} from "@/components/ui/card";
+import { RegisterForm } from "./page.client";
+
+const Page = () => {
+  return (
+    <div className="flex flex-col gap-6">
+      <Card className="border-0 shadow-none">
+        <CardHeader>
+          <CardTitle>Create your account</CardTitle>
+          <CardDescription>
+            Enter your details below to create your account
+          </CardDescription>
+        </CardHeader>
+        <CardContent>
+          <RegisterForm />
+        </CardContent>
+      </Card>
+    </div>
+  );
+};
+
+export default Page;
