--- src/app/auth/register/page.tsx
@@ -1,13 +1,0 @@
-import { RegisterForm } from "./page.client";
-
-const Page = () => {
-  return (
-    <div className="flex justify-center w-full">
-      <div className="w-full max-w-sm">
-        <RegisterForm />
-      </div>
-    </div>
-  );
-};
-
-export default Page;
